добавил room, поиск города по названию, изменил дизайн

+++ app/src/main/java/ru/maxdexter/myweather/model/WeatherData.kt
@@ -14,9 +14,9 @@ data class WeatherData(
     @SerializedName("hourly")
     val hourly: List<Hourly>,
     @SerializedName("lat")
-    val lat: Int,
+    val lat: Double,
     @SerializedName("lon")
-    val lon: Int,
+    val lon: Double,
     @SerializedName("minutely")
     val minutely: List<Minutely>,
     @SerializedName("timezone")
